+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 ----------
 - Introduced `--tag` option to various sub-commands to "tag" snapshots
   of the same subvolume for different backup destinations
+- Added support for working with remote systems via `--remote-command`
+  option
 - Fixed `--help` & `--version` options to no longer report errors